requirements.txt
# ============================================================
# STB Launcher Hybrid Automation Framework
# requirements.txt — Tested on Python 3.11.x (Windows x64)
# ============================================================
# IMPORTANT: Requires Python 3.11. Python 3.12+ is fine too.
# Python 3.13/3.14 NOT supported yet — numpy/pandas have no
# pre-built wheels for those versions as of early 2026.
#
# Install Python 3.11 from: https://www.python.org/downloads/release/python-31110/
# Direct link: https://www.python.org/ftp/python/3.11.10/python-3.11.10-amd64.exe
#
# Then create a venv with it:
#   py -3.11 -m venv .venv
#   .venv\Scripts\activate
#   pip install -r requirements.txt
# ============================================================

# Core Test Runner
pytest==8.3.5
pytest-xdist==3.6.1
pytest-rerunfailures==14.0
pytest-html==4.1.1
execnet==2.1.1

# Appium - UI Automation
Appium-Python-Client==4.3.0
selenium==4.26.0

# API Automation
requests==2.32.3
urllib3==2.2.3
certifi>=2024.2.2
charset-normalizer>=3.3.2
idna>=3.6

# Dashboard
streamlit==1.40.2
pandas==2.1.4
plotly==5.24.1
watchdog==4.0.1

# Streamlit auto-refresh
streamlit-autorefresh==1.0.1

# n8n Webhook Server
flask==3.0.3
flask-cors==4.0.1
waitress==3.0.0              # Production WSGI server for Windows (no eventlet needed)

# Reporting & Utilities
python-dotenv==1.0.1
colorlog==6.9.0
tenacity==9.0.0
Pillow==10.4.0
pytz==2024.2

# Optional: Allure reporting
allure-pytest==2.13.5

